+++ libs/core/algorithms/tests/unit/algorithms/destroy_tests.hpp
@@ -1,4 +1,4 @@
-//  Copyright (c) 2014-2020 Hartmut Kaiser
+//  Copyright (c) 2014-2025 Hartmut Kaiser
 //  Copyright (c) 2024 Tobias Wukovitsch
 //
 //  SPDX-License-Identifier: BSL-1.0
@@ -179,9 +179,9 @@ void test_destroy_exception(IteratorTag)
 
     HPX_TEST_EQ(data_type::instance_count.load(), data_size);
 
-    std::uniform_int_distribution<> dis(0, data_size - 1);
+    std::uniform_int_distribution<> dis(0, static_cast<int>(data_size - 1));
     std::atomic<std::size_t> throw_after(dis(gen));    //-V104
-    std::int64_t throw_after_ = throw_after.load();
+    std::size_t throw_after_ = throw_after.load();
 
     bool caught_exception = false;
     try
@@ -235,9 +235,9 @@ void test_destroy_exception(ExPolicy&& policy, IteratorTag)
 
     HPX_TEST_EQ(data_type::instance_count.load(), data_size);
 
-    std::uniform_int_distribution<> dis(0, data_size - 1);
+    std::uniform_int_distribution<> dis(0, static_cast<int>(data_size - 1));
     std::atomic<std::size_t> throw_after(dis(gen));    //-V104
-    std::int64_t throw_after_ = throw_after.load();
+    std::size_t throw_after_ = throw_after.load();
 
     bool caught_exception = false;
     try
@@ -288,9 +288,9 @@ void test_destroy_exception_async(ExPolicy&& policy, IteratorTag)
 
     HPX_TEST_EQ(data_type::instance_count.load(), data_size);
 
-    std::uniform_int_distribution<> dis(0, data_size - 1);
+    std::uniform_int_distribution<> dis(0, static_cast<int>(data_size - 1));
     std::atomic<std::size_t> throw_after(dis(gen));    //-V104
-    std::int64_t throw_after_ = throw_after.load();
+    std::size_t throw_after_ = throw_after.load();
 
     bool caught_exception = false;
     bool returned_from_algorithm = false;
@@ -351,9 +351,9 @@ void test_destroy_bad_alloc(ExPolicy&& policy, IteratorTag)
 
     HPX_TEST_EQ(data_type::instance_count.load(), data_size);
 
-    std::uniform_int_distribution<> dis(0, data_size - 1);
+    std::uniform_int_distribution<> dis(0, static_cast<int>(data_size - 1));
     std::atomic<std::size_t> throw_after(dis(gen));    //-V104
-    std::int64_t throw_after_ = throw_after.load();
+    std::size_t throw_after_ = throw_after.load();
 
     bool caught_bad_alloc = false;
     try
@@ -404,9 +404,9 @@ void test_destroy_bad_alloc_async(ExPolicy&& policy, IteratorTag)
 
     HPX_TEST_EQ(data_type::instance_count.load(), data_size);
 
-    std::uniform_int_distribution<> dis(0, data_size - 1);
+    std::uniform_int_distribution<> dis(0, static_cast<int>(data_size - 1));
     std::atomic<std::size_t> throw_after(dis(gen));    //-V104
-    std::int64_t throw_after_ = throw_after.load();
+    std::size_t throw_after_ = throw_after.load();
 
     bool caught_bad_alloc = false;
     bool returned_from_algorithm = false;
