+++ libs/full/resiliency_distributed/tests/performance/replay/1d_stencil_replay_distributed.cpp
@@ -75,7 +75,8 @@ public:
         for (std::size_t k = 0; k != subdomain_width + 1; ++k)
         {
             data_[k] = std::sin(2 * pi *
-                ((0.0 + subdomain_width * subdomain_index + k) /
+                ((0.0 + static_cast<double>(subdomain_width) * subdomain_index +
+                     static_cast<double>(k)) /
                     static_cast<double>(subdomain_width * subdomains)));
             checksum_ += data_[k];
         }
@@ -207,13 +208,16 @@ partition_data stencil_update(std::size_t sti, partition_data const& center,
     const std::size_t size = center.size() - 1;
     partition_data workspace(size + 2 * sti + 1);
 
-    std::copy(end(left) - sti - 1, end(left) - 1, &workspace[0]);
+    std::copy(end(left) - static_cast<std::ptrdiff_t>(sti) - 1, end(left) - 1,
+        &workspace[0]);
     std::copy(begin(center), end(center) - 1, &workspace[sti]);
-    std::copy(begin(right), begin(right) + sti + 1, &workspace[size + sti]);
+    std::copy(begin(right), begin(right) + static_cast<std::ptrdiff_t>(sti) + 1,
+        &workspace[size + sti]);
 
-    double left_checksum = std::accumulate(end(left) - sti - 1, end(left), 0.0);
-    double right_checksum =
-        std::accumulate(begin(right), begin(right) + sti + 1, 0.0);
+    double left_checksum = std::accumulate(
+        end(left) - static_cast<std::ptrdiff_t>(sti) - 1, end(left), 0.0);
+    double right_checksum = std::accumulate(
+        begin(right), begin(right) + static_cast<std::ptrdiff_t>(sti) + 1, 0.0);
 
     double checksum = left_checksum - center[0] + center.checksum() - right[0] +
         right_checksum;
